requirements.txt
torch  # recommend GPU version
opencv-python==4.9.0.80
mediapipe==0.10.11
TTS==0.22.0
pocketsphinx==5.0.3
pygame==2.5.2
websocket-client==1.7.0
urllib